1、修改PDF稳定性问题：PinchPhase手势缺少zoom属性，强转前未判空；2、调整PDF维测HandleInputEvent日志，输出事件类别为可阅读类型。

+++ pdf/pdfium/pdfium_engine.cc
@@ -1057,7 +1057,7 @@ bool PDFiumEngine::HandleInputEvent(const blink::WebInputEvent& event) {
   defer_page_unload_ = true;
   bool rv = false;
 #if BUILDFLAG(ARKWEB_PDF)
-  LOG(DEBUG) << "PDFiumEngine::HandleInputEvent " << static_cast<int>(event.GetType());
+  LOG(DEBUG) << "PDFiumEngine::HandleInputEvent " << event.GetType();
 #endif
   switch (event.GetType()) {
     case blink::WebInputEvent::Type::kMouseDown:
